Changed field value to be of type array
instead of object. This is to make Inputfield work as a config field
for modules where values are stored as serialized json array. You may
need to update your code if you set or change values via API somewhere
in your templates or modules. Look in read me for upgrade notes